Add some tests for CLI interface
